requirements.txt
boto3==1.34.34
python-dotenv==1.0.0
Pillow==10.0.0